// react/lib/schematic/api/checkoutexternal/models/CompanyDetailResponseData.ts
export interface CompanyDetailResponseData {
    /**
     * 
     * @type {Array<CompanyPlanWithBillingSubView>}
     * @memberof CompanyDetailResponseData
     */
    addOns: Array<CompanyPlanWithBillingSubView>;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof CompanyDetailResponseData
     */
    billingCreditBalances?: { [key: string]: number; } | null;
    /**
     * 
     * @type {BillingSubscriptionView}
     * @memberof CompanyDetailResponseData
     */
    billingSubscription?: BillingSubscriptionView;
    /**
     * 
     * @type {Array<BillingSubscriptionView>}
     * @memberof CompanyDetailResponseData
     */
    billingSubscriptions: Array<BillingSubscriptionView>;
    /**
     * 
     * @type {Date}
     * @memberof CompanyDetailResponseData
     */
    createdAt: Date;
    /**
     * 
     * @type {Array<CustomPlanBillingResponseData>}
     * @memberof CompanyDetailResponseData
     */
    customPlanBillings: Array<CustomPlanBillingResponseData>;
    /**
     * 
     * @type {PaymentMethodResponseData}
     * @memberof CompanyDetailResponseData
     */
    defaultPaymentMethod?: PaymentMethodResponseData;
    /**
     * 
     * @type {Array<FeatureEntitlement>}
     * @memberof CompanyDetailResponseData
     */
    entitlements: Array<FeatureEntitlement>;
    /**
     * 
     * @type {Array<EntityTraitDetailResponseData>}
     * @memberof CompanyDetailResponseData
     */
    entityTraits: Array<EntityTraitDetailResponseData>;
    /**
     * 
     * @type {string}
     * @memberof CompanyDetailResponseData
     */
    environmentId: string;
    /**
     * 
     * @type {string}
     * @memberof CompanyDetailResponseData
     */
    id: string;
    /**
     * 
     * @type {Array<EntityKeyDetailResponseData>}
     * @memberof CompanyDetailResponseData
     */
    keys: Array<EntityKeyDetailResponseData>;
    /**
     * 
     * @type {Date}
     * @memberof CompanyDetailResponseData
     */
    lastSeenAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyDetailResponseData
     */
    logoUrl?: string | null;
    /**
     * 
     * @type {Array<CompanyEventPeriodMetricsResponseData>}
     * @memberof CompanyDetailResponseData
     */
    metrics: Array<CompanyEventPeriodMetricsResponseData>;
    /**
     * 
     * @type {string}
     * @memberof CompanyDetailResponseData
     */
    name: string;
    /**
     * 
     * @type {Array<PaymentMethodResponseData>}
     * @memberof CompanyDetailResponseData
     */
    paymentMethods: Array<PaymentMethodResponseData>;
    /**
     * 
     * @type {CompanyPlanWithBillingSubView}
     * @memberof CompanyDetailResponseData
     */
    plan?: CompanyPlanWithBillingSubView;
    /**
     * 
     * @type {Array<GenericPreviewObject>}
     * @memberof CompanyDetailResponseData
     */
    plans: Array<GenericPreviewObject>;
    /**
     * 
     * @type {Array<Rule>}
     * @memberof CompanyDetailResponseData
     */
    rules: Array<Rule>;
    /**
     * 
     * @type {ScheduledDowngradeResponseData}
     * @memberof CompanyDetailResponseData
     */
    scheduledDowngrade?: ScheduledDowngradeResponseData;
    /**
     * A map of trait names to trait values
     * @type {object}
     * @memberof CompanyDetailResponseData
     */
    traits?: object;
    /**
     * 
     * @type {Date}
     * @memberof CompanyDetailResponseData
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof CompanyDetailResponseData
     */
    userCount: number;
}