// react/lib/schematic/api/checkoutexternal/models/ComponentHydrateResponseData.ts
export interface ComponentHydrateResponseData {
    /**
     * 
     * @type {Array<CompanyPlanDetailResponseData>}
     * @memberof ComponentHydrateResponseData
     */
    activeAddOns: Array<CompanyPlanDetailResponseData>;
    /**
     * 
     * @type {Array<CompanyPlanDetailResponseData>}
     * @memberof ComponentHydrateResponseData
     */
    activePlans: Array<CompanyPlanDetailResponseData>;
    /**
     * 
     * @type {Array<UsageBasedEntitlementResponseData>}
     * @memberof ComponentHydrateResponseData
     */
    activeUsageBasedEntitlements: Array<UsageBasedEntitlementResponseData>;
    /**
     * 
     * @type {Array<CompatiblePlans>}
     * @memberof ComponentHydrateResponseData
     */
    addOnCompatibilities: Array<CompatiblePlans>;
    /**
     * 
     * @type {ComponentCapabilities}
     * @memberof ComponentHydrateResponseData
     */
    capabilities?: ComponentCapabilities;
    /**
     * 
     * @type {ComponentCheckoutSettings}
     * @memberof ComponentHydrateResponseData
     */
    checkoutSettings: ComponentCheckoutSettings;
    /**
     * 
     * @type {CompanyDetailResponseData}
     * @memberof ComponentHydrateResponseData
     */
    company?: CompanyDetailResponseData;
    /**
     * 
     * @type {ComponentResponseData}
     * @memberof ComponentHydrateResponseData
     */
    component?: ComponentResponseData;
    /**
     * 
     * @type {Array<BillingCreditBundleView>}
     * @memberof ComponentHydrateResponseData
     */
    creditBundles: Array<BillingCreditBundleView>;
    /**
     * 
     * @type {Array<CreditCompanyGrantView>}
     * @memberof ComponentHydrateResponseData
     */
    creditGrants: Array<CreditCompanyGrantView>;
    /**
     * 
     * @type {Array<CheckoutFieldWithValue>}
     * @memberof ComponentHydrateResponseData
     */
    customCheckoutFields: Array<CheckoutFieldWithValue>;
    /**
     * 
     * @type {PlanDetailResponseData}
     * @memberof ComponentHydrateResponseData
     */
    defaultPlan?: PlanDetailResponseData;
    /**
     * 
     * @type {ComponentDisplaySettings}
     * @memberof ComponentHydrateResponseData
     */
    displaySettings: ComponentDisplaySettings;
    /**
     * 
     * @type {FeatureUsageDetailResponseData}
     * @memberof ComponentHydrateResponseData
     */
    featureUsage?: FeatureUsageDetailResponseData;
    /**
     * 
     * @type {PlanDetailResponseData}
     * @memberof ComponentHydrateResponseData
     */
    postTrialPlan?: PlanDetailResponseData;
    /**
     * 
     * @type {boolean}
     * @memberof ComponentHydrateResponseData
     */
    preventSelfServiceDowngrade: boolean;
    /**
     * 
     * @type {string}
     * @memberof ComponentHydrateResponseData
     */
    preventSelfServiceDowngradeButtonText?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComponentHydrateResponseData
     */
    preventSelfServiceDowngradeUrl?: string | null;
    /**
     * 
     * @type {ScheduledDowngradeResponseData}
     * @memberof ComponentHydrateResponseData
     */
    scheduledDowngrade?: ScheduledDowngradeResponseData;
    /**
     * 
     * @type {boolean}
     * @memberof ComponentHydrateResponseData
     * @deprecated
     */
    showAsMonthlyPrices: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ComponentHydrateResponseData
     * @deprecated
     */
    showCredits: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ComponentHydrateResponseData
     * @deprecated
     */
    showPeriodToggle: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ComponentHydrateResponseData
     * @deprecated
     */
    showZeroPriceAsFree: boolean;
    /**
     * 
     * @type {StripeEmbedInfo}
     * @memberof ComponentHydrateResponseData
     */
    stripeEmbed?: StripeEmbedInfo;
    /**
     * 
     * @type {CompanySubscriptionResponseData}
     * @memberof ComponentHydrateResponseData
     */
    subscription?: CompanySubscriptionResponseData;
    /**
     * 
     * @type {boolean}
     * @memberof ComponentHydrateResponseData
     */
    trialPaymentMethodRequired?: boolean | null;
    /**
     * 
     * @type {InvoiceResponseData}
     * @memberof ComponentHydrateResponseData
     */
    upcomingInvoice?: InvoiceResponseData;
}